fix(char)!: Move `escaped` from `bytes` to `character`
